Update localTree.js
The change refers to fireftp Issue #64 - "File copy and paste within same directory doesn't work (Local as well as remote)".


The idea is to do copy and paste of a file in the same directory. The new file gets the extension _(x) instead of overwriting the same file.

I don't have javascript development environment. So the changes above are beta and need to be tested. File and path names must be corrected. For an experienced javascript developer this should be easy to do. The approach is already done by myself.